Extract ConfigurationCacheLambdaIntegrationTest
from ConfigurationCacheIntegrationTest
