+++ subprojects/configuration-cache/src/integTest/groovy/org/gradle/configurationcache/ConfigurationCacheIntegrationTest.groovy
@@ -772,112 +772,6 @@ class ConfigurationCacheIntegrationTest extends AbstractConfigurationCacheIntegr
     }
 
     @Unroll
-    def "restores task fields whose value is a serializable #kind Java lambda"() {
-        given:
-        file("buildSrc/src/main/java/my/LambdaTask.java").tap {
-            parentFile.mkdirs()
-            text = """
-                package my;
-
-                import org.gradle.api.*;
-                import org.gradle.api.tasks.*;
-
-                public class LambdaTask extends DefaultTask {
-
-                    public interface SerializableSupplier<T> extends java.io.Serializable {
-                        T get();
-                    }
-
-                    private SerializableSupplier<Integer> supplier;
-
-                    public void setSupplier(SerializableSupplier<Integer> supplier) {
-                        this.supplier = supplier;
-                    }
-
-                    public void setNonInstanceCapturingLambda() {
-                        final int i = getName().length();
-                        setSupplier(() -> i);
-                    }
-
-                    public void setInstanceCapturingLambda() {
-                        setSupplier(() -> getName().length());
-                    }
-
-                    @TaskAction
-                    void printValue() {
-                        System.out.println("this.supplier.get() -> " + this.supplier.get());
-                    }
-                }
-            """
-        }
-
-        buildFile << """
-            task ok(type: my.LambdaTask) {
-                $expression
-            }
-        """
-
-        when:
-        configurationCacheRun "ok"
-        configurationCacheRun "ok"
-
-        then:
-        outputContains("this.supplier.get() -> 2")
-
-        where:
-        kind                     | expression
-        "instance capturing"     | "setInstanceCapturingLambda()"
-        "non-instance capturing" | "setNonInstanceCapturingLambda()"
-    }
-
-    @Unroll
-    def "restores task with action and spec that are Java lambdas"() {
-        given:
-        file("buildSrc/src/main/java/my/LambdaPlugin.java").tap {
-            parentFile.mkdirs()
-            text = """
-                package my;
-
-                import org.gradle.api.*;
-                import org.gradle.api.tasks.*;
-
-                public class LambdaPlugin implements Plugin<Project> {
-                    public void apply(Project project) {
-                        $type value = $expression;
-                        project.getTasks().register("ok", task -> {
-                            task.doLast(t -> {
-                                System.out.println(task.getName() + " action value is " + value);
-                            });
-                            task.onlyIf(t -> {
-                                System.out.println(task.getName() + " spec value is " + value);
-                                return true;
-                            });
-                        });
-                    }
-                }
-            """
-        }
-
-        buildFile << """
-            apply plugin: my.LambdaPlugin
-        """
-
-        when:
-        configurationCacheRun "ok"
-        configurationCacheRun "ok"
-
-        then:
-        outputContains("ok action value is ${value}")
-        outputContains("ok spec value is ${value}")
-
-        where:
-        type      | expression | value
-        "String"  | '"value"'  | "value"
-        "int"     | "12"       | "12"
-        "boolean" | "true"     | "true"
-    }
-
-    @Unroll
     def "restores task fields whose value is #kind TextResource"() {
 
         given:
